refactor: split store into modules
